Correctly typed model collection
